automation/monitoring/requirements.txt
functions-framework==3.*
google-cloud-monitoring
google-cloud-firestore
pandas
scikit-learn
numpy